+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- `isInteger` validation to `number` type form elements
+
 ## 0.1.3 (2020-11-05)
 
 ### Added